Gave Client its own tcp/udp code rather than using SharedClassLibrarys.  This means low level network code can be seperate for client and server.

+++ Chicken VS Chicken Game Server/src/Server/Networking/Client.cs
@@ -70,7 +70,7 @@ namespace GameServer
         public void HandlePacket(Packet _packet)
         {
             // TODO: change client packet id type to byte
-            int _packetId = _packet.ReadInt();
+            byte _packetId = _packet.ReadByte();
             packetHandlers[_packetId](this, _packet);
         }
